NotebookToLatex.jl

Why this Package?

This package converts your notebook files (Pluto or Jupyter) to beautiful and simple Latex files, that are easy to modify. Thus, making it ideal to write reports, articles or books from notebooks.

Although it's already possible to convert both Pluto and Jupyter notebooks to PDFs, or even to Latex (via Pandoc), the output Latex files are usually very messy. In contrast, NotebookToLatex.jl is very opinionated and specific, producing straightforward latex files.

This package uses julia-mono-listing, to produce beautiful Julia code inside the Latex pdf. Note that it requires using lualatex for compilation.

Getting Started

This package is very simple to use. There is pretty much just one function to be used, i.e.

NotebookToLatex.notebooktolatex — Function

notebooktolatex(notebook, targetdir="./build_latex"; template=:book, fontpath=nothing)

Takes a notebook file, converts it to Latex and creates a file structure with figures, fonts and listing files.

targetdir is the target directory where the Latex project will be created.

If the directory does no exists, it is created.

template - The template for the Latex file. It's based on Latex templates.

Current supported templates are :book, :mathbook.

fontpath - The output Latex files uses JuliaMono fonts in order to support the

unicodes that are also supported in Julia. If the user already has JuliaMono installed, he can provide the path to where the .ttf files are stored. If nothing is passed, then the font files will be downloaded and saved in the ./fonts/ folder.

Basic Use

To convert the notebooks just use notebooktolatex("mynotebook.jl", template=:book). This will produce a directory ./build_latex/ where the Latex files will be generated. Inside build_latex/ you will have:

build_latex
│   main.tex
│   julia_font.tex
│   julia_listings.tex
│   julia_listings_unicode.tex
│   preface.tex
│
└───figures
│   │   mynotebook_plot1.png
│   └───mynotebook_plot2.png
└───fonts
│   │   JuliaMono_Regular.ttf
│   │   ...
│   
└───frontmatter
│   │   titlepage.tex
│   └───copyright.tex
│
└───notebooks
    └───mynotebook.tex

Using template=:book, we get the Latex book format, thus, we have a preface.tex, a titlepage.tex and a copyright.tex page. The notebook will be included as a chapter. To get your final book pdf, just compile the main.tex using lualatex.

At the moment, the available templates are:

:book - The standard Latex book template;

:mathbook - Very similar to :book, but with some extra packages already imported.